Macro escaping: Escape macros in command lines as macros should
This wasn't a problem, due to a bug in nagios macro escaping, so lot of macros
was escaped anyway, so macros wouldn't break the command line anyway.

When this bug is fixed, it's actually up to the requestor to request a command
line escaped string, so lets do that too.

This only affects upcoming versions of nagios 4, but shouldn't break anything
in versions prior to that.

Signed-off-by: Max Sikström <msikstrom@op5.com>